refactor: header 폴더 생성
